Update: Updated dash_attack
Updated the dash_attack animation to be special_side with special_side being replaced at a later date. Made the move slightly slower and cleaned up misc blank keyframes. Also adjusted the xSpeedConservation to have the move get a small boost of momentum when using it. Hitboxes were also adjusted to not kill as early and have hitStopOffset.